Create IntroCard component
