+++ src/components/IntroCard/index.ts
@@ -1,0 +1,1 @@
+export * from './IntroCard'
